Add a perf backed PerfTrace class
Adds a PerfTrace class with similar semantics to the GpuTrace class that
uses perf rather than trace-cmd to generate trace data.

In a future commit, this class will be integrated into all gpu-trace
modes such that it will be possible to generate a trace-cmd trace and a
perf trace simultaneously, as well as open both of these traces in
gpuvis. Until this commit, the only way to use this code is by importing
gpu-trace.py as a module and instantiating PerfTrace manually.

This can be done as follows:

    >>> import importlib as imp
    >>> gt = imp.import_module("gpu-trace")
    >>> ct = gt.PerfTrace()

The following lines can be used to enable logging:

    >>> import logging
    >>> gt.SetupLogging("/tmp/gt.log", logging.DEBUG)

From this state, it is possible to start recording, stop recording,
and capture a trace as follows:

    >>> ct.StartRecord()
    >>> ct.CaptureTrace("/tmp/trace.json")
    >>> ct.StopRecord()

Signed-off-by: Shawn M. Chapla <schapla@codeweavers.com>