Trim ftrace traces to a maximum duration
First, find a good reference time for end of capture in the trace-cmd
output - we pick the last sched_exec of the trace-cmd executable, as
this is almost certainly our 'trace-cmd stop' invocation.

Then trim the trace-cmd output using that end time and a command
line specified duration.

For now, perf results are still full buffer.